better search with section headings
